+++ jmespath/__init__.py
@@ -1,7 +1,7 @@
 from jmespath import parser
 from jmespath.visitor import Options
 
-__version__ = '0.9.2'
+__version__ = '0.9.3'
 
 
 def compile(expression):