+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+0.9.3
+=====
+
+* Fix issue where long types in py2 and ``Decimal`` types
+  were not being evaluated as numbers
+  (`issue 125 <https://github.com/jmespath/jmespath.py/issues/125>`__)
+* Handle numbers in scientific notation in ``to_number()`` function
+  (`issue 120 <https://github.com/jmespath/jmespath.py/issues/120>`__)
+* Fix issue where custom functions would override the function table
+  of the builtin function class
+  (`issue 133 <https://github.com/jmespath/jmespath.py/issues/133>`__)
+
+
 0.9.2
 =====